Extract tests for isSymbol to own file
